refactor(api): knowledge → src/api modules (Effort 1)
Mechanical move of the knowledge feature from src/hooks/queries/knowledge to
src/api/knowledge, adding queryOptions factories for the query hooks. No behavior change.
