+++ src/api/knowledge/types.ts
@@ -1,0 +1,4 @@
+export const knowledgeKeys = {
+  all: ["knowledge"] as const,
+  user: (userId: string) => [...knowledgeKeys.all, "user", userId] as const,
+};
